Append KeyFrame time measuring..

+++ lsd_slam_viewer/src/PointCloudViewer.cpp
@@ -327,8 +327,8 @@ void PointCloudViewer::keyPressEvent(QKeyEvent *e)
     	  meddleMutex.lock();
 
 
-  //      qreal x,y,z;
-          float x,y,z;
+      qreal x,y,z;
+//          float x,y,z;
     	  camera()->frame()->getPosition(x,y,z);
     	  animationList.push_back(AnimationObject(false, lastAnimTime, 2, qglviewer::Frame(qglviewer::Vec(0,0,0), camera()->frame()->orientation())));
     	  animationList.back().frame.setPosition(x,y,z);
